apps/user-service/requirements.txt
SQLAlchemy==2.0.23
werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0
alembic==1.12.0
python-jose[cryptography]==3.3.0
fastapi==0.103.2
uvicorn[standard]==0.23.2
httpx==0.25.0
pydantic-settings==2.1.0
python-multipart==0.0.6